move func to utils
